Bug 1520458 - Update Raptor to run on python 3 r=perftest-reviewers,davehunt
Differential Revision: https://phabricator.services.mozilla.com/D77315

UltraBlame original commit: 76bffd1b3c6ecdba621930c770c52aae6e6205cc

+++ testing/raptor/raptor/filters.py
@@ -121,7 +121,7 @@ def median(series):
         return series[len(series)/2]
     else:
         
-        middle = len(series)/2  
+        middle = int(len(series)/2)  
         return 0.5*(series[middle-1] + series[middle])
